requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
scikit-learn
pypdf
python-multipart
requests
passlib
bcrypt==4.0.1
python-jose
httpx
pandas
streamlit
